Partial style rewrite.

+++ kierki-serwer.cpp
@@ -3,8 +3,8 @@
 #include <exception>
 #include <string>
 
-#include "serwer.h"
 #include "cmd_args_parsers.h"
+#include "serwer.h"
 #include "file_reader.h"
 
 using std::cout;
@@ -16,11 +16,12 @@ using std::invalid_argument;
 
 int main(int argc, char* argv[])
 {
-    int port = 0;
+    int32_t port = 0;
     string game_file_name;
-    int timeout = 5;
+    int32_t timeout = 5;
     
-    int result = parser::parse_server_args(argc, argv, port, game_file_name, timeout);
+    int16_t result = parser::parse_server_args(argc, argv, port,
+        game_file_name, timeout);
     if (result != 0) {return result;}
 
     Serwer s(port, timeout, game_file_name);
@@ -28,5 +29,3 @@ int main(int argc, char* argv[])
     else {return 1;}
     return 0;
 }
-
-// g++ -I/home/gustaw/boost_library/boost_1_74_0 -std=c++20  kierki-serwer.cpp -o kierki -L/home/gustaw/boost_library/boost_1_74_0/stage/lib -lm -l:libboost_program_options.a
